+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 
 ## master
 
-- use [RGBDS:8b60efa149](https://github.com/rednex/rgbds/tree/8b60efa149)
-  *(2020-06-27)*
 - use [dmg-acid2:8a98ce731f](https://github.com/mattcurrie/dmg-acid2/tree/8a98ce731f)
   *(2020-04-21)*
+- use [cgb-acid2:76c4e2c42d](https://github.com/mattcurrie/cgb-acid2/tree/76c4e2c42d)
+  *(2020-04-21)*
+- compile dmg-acid2 and cgb-acid2 test roms with
+  [RGBDS:8b60efa149](https://github.com/rednex/rgbds/tree/8b60efa149)
+  *(2020-06-27)*
 
 
 ## [v1.1](https://github.com/c-sp/gameboy-test-roms/releases/tag/v1.1) *(2020-06-14)*